fix: use BIGSERIAL for ID columns to match JPA Long type

+++ src/main/resources/schema.sql
@@ -2,7 +2,7 @@
 -- PostgreSQL
 
 CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     uid VARCHAR(255) UNIQUE NOT NULL,
     email VARCHAR(100) UNIQUE NOT NULL,
     first_name VARCHAR(100),
@@ -13,16 +13,16 @@ CREATE TABLE users (
 );
 
 CREATE TABLE stats (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     xp INTEGER NOT NULL DEFAULT 0,
     games_played INTEGER NOT NULL DEFAULT 0,
     questions_correct INTEGER NOT NULL DEFAULT 0,
     questions_wrong INTEGER NOT NULL DEFAULT 0,
-    user_id INTEGER NOT NULL UNIQUE REFERENCES users(id) ON DELETE CASCADE
+    user_id BIGINT NOT NULL UNIQUE REFERENCES users(id) ON DELETE CASCADE
 );
 
 CREATE TABLE countries (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     flag TEXT NOT NULL,
     country_code VARCHAR(2),
     name VARCHAR(30),
@@ -31,19 +31,19 @@ CREATE TABLE countries (
 );
 
 CREATE TABLE case_files (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     article_id INTEGER UNIQUE,
     article_content TEXT,
     article_title TEXT,
     publish_date VARCHAR(50),
     summary_young TEXT DEFAULT NULL,
     summary_old TEXT DEFAULT NULL,
-    countries_id INTEGER REFERENCES countries(id),
+    countries_id BIGINT REFERENCES countries(id),
     photo_url TEXT
 );
 
 CREATE TABLE questions_younger (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     y_question VARCHAR(150) NOT NULL,
     y_correct_answer VARCHAR(100) NOT NULL,
     y_incorrect_answer1 VARCHAR(100) NOT NULL,
@@ -53,7 +53,7 @@ CREATE TABLE questions_younger (
 );
 
 CREATE TABLE questions_older (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     o_question VARCHAR(150) NOT NULL,
     o_correct_answer VARCHAR(100) NOT NULL,
     o_incorrect_answer1 VARCHAR(100) NOT NULL,
